Split out helpers to remove the need for openseadragon mocking
