+++ content/webapp/views/pages/works/work/IIIFItem/index.tsx
@@ -52,7 +52,7 @@ import BetaMessage from '@weco/content/views/components/BetaMessage';
 import { toWorksItemLink } from '@weco/content/views/components/ItemLink';
 import VideoPlayer from '@weco/content/views/components/VideoPlayer';
 import IIIFItemPdf from '@weco/content/views/pages/works/work/IIIFItem/IIIFItem.Pdf';
-import { arrayIndexToQueryParam } from '@weco/content/views/pages/works/work/IIIFViewer';
+import { arrayIndexToQueryParam } from '@weco/content/views/pages/works/work/IIIFViewer/IIIFViewer.helpers';
 import ImageViewer from '@weco/content/views/pages/works/work/IIIFViewer/ImageViewer';
 import RestrictedItemMessage from '@weco/content/views/pages/works/work/work.RestrictedItemMessage';
