+++ content/webapp/views/pages/works/work/IIIFViewer/ViewerStructures.tsx
@@ -21,7 +21,7 @@ import {
 } from '@weco/content/utils/iiif/v3';
 import { toWorksItemLink } from '@weco/content/views/components/ItemLink';
 
-import { arrayIndexToQueryParam } from '.';
+import { arrayIndexToQueryParam } from './IIIFViewer.helpers';
 import { thumbnailsPageSize } from './Paginators';
 
 export const List = styled(PlainList)`
