+++ content/webapp/views/pages/works/work/IIIFViewer/ViewerTopBar.test.tsx
@@ -14,11 +14,6 @@ import { TransformedManifest } from '@weco/content/types/manifest';
 
 import ViewerTopBar from './ViewerTopBar';
 
-// ViewerTopBar transitively imports the whole viewer (via the `.` barrel that
-// re-exports queryParamToArrayIndex), which pulls in openseadragon. It is never
-// exercised by this component, so stub it out to avoid the jsdom canvas error.
-jest.mock('openseadragon', () => ({ __esModule: true, default: jest.fn() }));
-
 // A manifest-level rendering that yields a non-empty downloadOptions list.
 const pdfRendering: TransformedManifest['rendering'] = [
   {
